pyproject.toml
[project]
name = "picgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "augraphy>=8.2.6",
    "mmcv",
    "mmengine",
    "mmocr>=1.0.1",
    "numpy>=2.2.6",
    "openai>=1.97.1",
    "paddleocr>=3.1.0",
    "paddlepaddle==3.1.0",
    "pillow>=11.3.0",
    "ppocrlabel>=3.1.2",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
]

[tool.uv.sources]
mmengine = { git = "https://github.com/open-mmlab/mmengine.git" }
mmcv = { git = "https://gitcode.com/open-mmlab/mmcv.git" }

[[tool.uv.index]]
url = "https://mirrors.ustc.edu.cn/pypi/simple"
default = true
